- Update RJSF to 6.0.2 - Refactor `BatchLayout` to `SecondaryNavigation` for re-use - Elevate BatchLayout out of Suspense for better layout on navigation - Disentangle Admin page into 4 pages, extracting common functionality - Add navigation for that - Fix selectionModel on MUI grids - Fix deprecated `enumNames` in API spec and test
