+++ app/package.json
@@ -27,10 +27,10 @@
     "@mui/x-data-grid": "^8.14.1",
     "@mui/x-date-pickers": "^8.14.1",
     "@reduxjs/toolkit": "^2.9.1",
-    "@rjsf/core": "^6.0.0-beta.21",
-    "@rjsf/mui": "^6.0.0-beta.21",
-    "@rjsf/utils": "^6.0.0-beta.21",
-    "@rjsf/validator-ajv8": "^6.0.0-beta.21",
+    "@rjsf/core": "^6.0.2",
+    "@rjsf/mui": "^6.0.2",
+    "@rjsf/utils": "^6.0.2",
+    "@rjsf/validator-ajv8": "^6.0.2",
     "@testing-library/react": "16.3.0",
     "@testing-library/user-event": "14.6.1",
     "@turf/area": "^7.2.0",
