+++ app/src/UI/Features/Batch/Batch.css
@@ -1,23 +1,3 @@
-.batchNav {
-  background: #1a5a96;
-  padding: 0;
-  margin: 0;
-  padding-bottom: 1px;
-
-  a {
-    color: white !important;
-    text-decoration: none;
-  }
-
-  .current_batch_link {
-    border-bottom: 4px solid #fcba19;
-  }
-}
-
-.batchDarkNav {
-  background: #272727;
-}
-
 div.template-description {
   text-align: center;
   border: 1px solid #444;
