+++ app/src/UI/Features/Batch/BatchCreateNew.tsx
@@ -1,12 +1,7 @@
 import BatchCreate from 'UI/Features/Batch/batch-upload/BatchCreate';
-import BatchLayout from 'UI/Features/Batch/BatchLayout';
 
 const BatchCreateNew = () => {
-  return (
-    <BatchLayout>
-      <BatchCreate />
-    </BatchLayout>
-  );
+  return <BatchCreate />;
 };
 
 export default BatchCreateNew;
