+++ app/src/UI/Features/Batch/BatchList.tsx
@@ -1,14 +1,11 @@
 import Container from '@mui/material/Container';
-import BatchLayout from 'UI/Features/Batch/BatchLayout';
 import BatchUploadList from 'UI/Features/Batch/batch-upload/BatchUploadList';
 
 const BatchList = () => {
   return (
-    <BatchLayout>
-      <Container maxWidth={'lg'}>
-        <BatchUploadList />
-      </Container>
-    </BatchLayout>
+    <Container maxWidth={'lg'}>
+      <BatchUploadList />
+    </Container>
   );
 };
