+++ app/src/UI/Features/Batch/BatchTemplates.tsx
@@ -1,12 +1,7 @@
 import TemplateDownloadList from 'UI/Features/Batch/batch-upload/TemplateDownloadList';
-import BatchLayout from 'UI/Features/Batch/BatchLayout';
 
 const BatchTemplates = () => {
-  return (
-    <BatchLayout>
-      <TemplateDownloadList />
-    </BatchLayout>
-  );
+  return <TemplateDownloadList />;
 };
 
 export default BatchTemplates;
