+++ app/src/UI/Features/Batch/BatchView.tsx
@@ -1,15 +1,10 @@
 import BatchDetail from 'UI/Features/Batch/batch-upload/BatchDetail';
-import BatchLayout from 'UI/Features/Batch/BatchLayout';
 import { useParams } from 'react-router';
 
 const BatchView = () => {
   const { id } = useParams<{ id: string }>();
 
-  return (
-    <BatchLayout>
-      <BatchDetail id={id} />
-    </BatchLayout>
-  );
+  return <BatchDetail id={id} />;
 };
 
 export default BatchView;
